implementacion de servicios y controller

+++ src/main/java/pe/edu/cibertec/apprueditas/back_end/dto/SearchResponseDTO.java
@@ -7,6 +7,6 @@ public record SearchResponseDTO(String codigo,
                                 String marca,
                                 String modelo,
                                 Integer nroAsientos,
-                                BigDecimal precio,
+                                Integer precio,
                                 String color) {
 }
